Imported JS calls don't handle differing arities.
Right now there is a spec assert that host call's iterable result and wasm have the same arity. There's no guarantee this is true. Instead we should probably throw a type error if they don't match. Alternatively, we could do the more JS-y approach and allow the arity of result to be at least the arity of the expected signature.

I also normalized the wording around what happens when an exception is thrown in the host call does to the surrounding Wasm code. Previously , the wording was a little vague.